refacto(fortify/cli): split fortify ... command into fortify chat ... and fortify serve ...
