+++ README.md
@@ -39,7 +39,7 @@ If you just want to install `fortify` and try the terminal chat:
 ```bash
 python -m pip install -e .
 cp .env.sample .env
-fortify --agent example_agent
+fortify chat --agent example_agent
 ```
 
 Required keys for the example CLI flow:
@@ -51,10 +51,10 @@ Required keys for the example CLI flow:
 Useful next commands:
 
 ```bash
-fortify --list-agents
-fortify --agent researcher
-fortify --use examples/file_agents.py --agent workspace_explorer
-fortify --use examples/research_agents.py --agent update_researcher
+fortify chat --list-agents
+fortify chat --agent researcher
+fortify chat --use examples/file_agents.py --agent workspace_explorer
+fortify chat --use examples/research_agents.py --agent update_researcher
 ```
 
 The included local agent lives in `examples/example_agent/`, and the CLI can also load:
@@ -83,12 +83,12 @@ pnpm dev
 # 3. Add to asianf/.env:
 #        FORTIFY_KEY=fty_test_support-bot_...
 # 4. Start serve mode:
-uv run fortify --serve
+uv run fortify serve
 ```
 
 Then open http://localhost:5173/playground — type a message, watch the live stream of tool calls and policy decisions from your local agent.
 
-The dashboard's `/agents` page lets you edit each agent's YAML and policy. `fortify --serve` re-fetches at every turn boundary, so your edits take effect on the next chat message without a restart.
+The dashboard's `/agents` page lets you edit each agent's YAML and policy. `fortify serve` re-fetches at every turn boundary, so your edits take effect on the next chat message without a restart.
 
 ## ✨ Core Primitives
 
@@ -456,9 +456,9 @@ It demonstrates:
 For the CLI, you can import that script and then pick one of its registered agents:
 
 ```bash
-fortify --use examples/file_agents.py --agent workspace_explorer
-fortify --use examples/file_agents.py --agent repo_editor
-fortify --use examples/research_agents.py --agent update_researcher
+fortify chat --use examples/file_agents.py --agent workspace_explorer
+fortify chat --use examples/file_agents.py --agent repo_editor
+fortify chat --use examples/research_agents.py --agent update_researcher
 ```
 
 ## 🗂️ Builtin And Local Agents
@@ -486,7 +486,7 @@ The CLI also discovers local agents from:
 This repo ships a demo agent at `examples/example_agent/`, so from the project root you can simply run:
 
 ```bash
-fortify --agent example_agent
+fortify chat --agent example_agent
 ```
 
 ## 🔐 Policy Shape
@@ -835,22 +835,22 @@ python examples/demo.py
 Run the inline chat CLI with a local or builtin YAML agent:
 
 ```bash
-fortify --agent example_agent
+fortify chat --agent example_agent
 ```
 
 Run the CLI with code-defined agents from a Python script:
 
 ```bash
-fortify --use examples/file_agents.py --agent workspace_explorer
-fortify --use examples/file_agents.py --agent repo_editor
-fortify --use examples/research_agents.py --agent update_researcher
-fortify --use examples/research_agents.py --agent update_researcher --approval-mode ask
+fortify chat --use examples/file_agents.py --agent workspace_explorer
+fortify chat --use examples/file_agents.py --agent repo_editor
+fortify chat --use examples/research_agents.py --agent update_researcher
+fortify chat --use examples/research_agents.py --agent update_researcher --approval-mode ask
 ```
 
 List what the CLI can currently resolve:
 
 ```bash
-fortify --list-agents
+fortify chat --list-agents
 ```
 
 Register a code-defined agent's manifest with the Fortify platform. `--agent`
@@ -901,7 +901,7 @@ Endpoints:
 - `GET /v1/projects/:id/agents` — list agents with their YAMLs
 - `GET /v1/projects/:id/agents/:name` — read one agent
 - `PUT /v1/projects/:id/agents/:name` — save agent / policy / system YAMLs
-- `WS /v1/projects/:id/serve` — producer socket (the `--serve` CLI dials here)
+- `WS /v1/projects/:id/serve` — producer socket (the `fortify serve` CLI dials here)
 - `WS /v1/projects/:id/chat` — consumer socket (the dashboard Playground dials here)
 
 DB lives at `platform/api/fortify.db`. Delete it and restart to wipe state.
@@ -927,7 +927,7 @@ Routes:
 
 The dev server proxies `/v1/*` (HTTP and WebSocket) to `localhost:8000`, so HMR works through the same origin as the API.
 
-### Serve Mode (`fortify --serve`)
+### Serve Mode (`fortify serve`)
 
 Bridges your local agent runtime to the dashboard via the platform's WebSocket relay — same pattern as Cloudflare Tunnel or ngrok.
 
@@ -939,7 +939,7 @@ FORTIFY_PROJECT_ID=support-bot              # optional, parsed from key prefix
 FORTIFY_API_URL=http://localhost:8000       # optional, defaults to localhost:8000
 
 # run
-uv run fortify --serve
+uv run fortify serve
 ```
 
 Behaviour:
@@ -954,8 +954,8 @@ Behaviour:
 To override the agent at the CLI:
 
 ```bash
-fortify --serve --agent read_only
-fortify --serve --use examples/file_agents.py --agent workspace_explorer
+fortify serve --agent read_only
+fortify serve --use examples/file_agents.py --agent workspace_explorer
 ```
 
 ### How `load_agent()` resolves with `FORTIFY_KEY`
