+++ fortify/cli/register/__init__.py
@@ -3,6 +3,6 @@
 from fortify.cli.register.register import register_agent
 from fortify.cli.register.manifest import create_manifest
 from fortify.cli.register.models import AgentManifest
-from fortify.cli.register.main import main
+from fortify.cli.register.main import add_parser, main
 
-__all__ = ["register_agent", "create_manifest", "AgentManifest", "main"]
+__all__ = ["register_agent", "create_manifest", "AgentManifest", "add_parser", "main"]
